requirements.txt





clarkpycore
flowshow
python-dotenv
loguru
pytest
pyyaml

# mlflow